refactor(agent_loop): split loop.py into tool_dispatch/safety/stream_retry/events siblings
``backend/app/core/agent_loop/loop.py`` was 827 lines — over the
500-line file budget. Split into role-named sibling modules inside the
existing ``app/core/agent_loop/`` package; ``loop.py`` becomes a thin
193-line public entry that delegates to the new modules.

Module layout:
- ``loop.py`` (193 L) — ``agent_loop`` public entry + ``_run_loop``
  orchestrator only. Imports the helpers from the four new modules.
- ``tool_dispatch.py`` (204 L) — ``_execute_and_log_tool_call``,
  ``_dispatch_tool_call``, ``_check_permission_or_none``,
  ``_truncate_for_log`` (log helper).
- ``safety.py`` (188 L) — pre-turn iteration / wall-clock checks
  (``_check_iteration_safety``), tool-result collection
  (``_collect_tool_results``), the after-turn ``_should_stop``
  predicate, ``_terminated``, ``_make_context_snapshot``.
- ``stream_retry.py`` (269 L) — ``_StreamOutcome`` /
  ``_StreamAttemptState`` dataclasses + the full retry-with-backoff
  ladder (``_stream_with_retry``, ``_stream_attempt_events``,
  ``_apply_done_event``, ``_retry_budget_exhausted``,
  ``_copy_stream_outcome``, ``_stream_interrupted_after_events``,
  ``_sleep_before_retry``).
- ``events.py`` (119 L) — provider ``LLMEvent`` →
  ``AgentEvent`` translation (``_consume_llm_event``,
  ``_carry_provider_state``) + per-turn context preparation
  (``_prepare_llm_messages``).

The public surface of ``app.core.agent_loop`` is unchanged —
``agent_loop`` is still re-exported by ``app/core/agent_loop/__init__.py``;
tests that ``from app.core.agent_loop.loop import agent_loop`` keep
working because ``agent_loop`` is still in ``loop.py``.

Cross-module imports inside ``agent_loop/`` use absolute paths
(``from app.core.agent_loop.events import …``) per ruff TID252.

Removed the ``backend/app/core/agent_loop/loop.py`` entry from
``scripts/check-file-lines.mjs`` ``EXEMPT_PATH_FRAGMENTS``. Updated the
surrounding comment to note ``claude_provider.py`` + ``telegram/bot.py``
are the last two cct-stack offenders.

Validation:
- ``node scripts/check-file-lines.mjs`` → OK (no files exceed 500).
- ``cd backend && uv run mypy`` → 0 errors across 344 source files.
- ``cd backend && uv run ruff check .`` → clean.
- ``cd backend && uv run ruff format --check .`` → 345 files already formatted.
- ``cd backend && uv run lint-imports --config .importlinter`` →
  3 contracts kept, 0 broken.
- ``pytest test_agent_loop{,_scenarios,_safety}.py`` → 24 passed.
