+++ backend/app/core/agent_loop/loop.py
@@ -13,18 +13,34 @@ The loop owns:
 
 Each provider supplies a StreamFn — the only provider-specific code.
 The loop never imports any provider SDK directly.
+
+This module is the entry point. Per-concern helpers live in sibling
+modules within the :mod:`app.core.agent_loop` package:
+
+* :mod:`app.core.agent_loop.tool_dispatch` — permission gate + tool
+  invocation + per-call observability.
+* :mod:`app.core.agent_loop.safety` — pre-turn iteration / wall-clock
+  checks, tool-result collection, and the after-turn stop predicate.
+* :mod:`app.core.agent_loop.events` — translation of one provider
+  ``LLMEvent`` into one or more ``AgentEvent`` items + per-turn context
+  preparation.
+* :mod:`app.core.agent_loop.stream_retry` — provider stream retry budget
+  and partial-output handling.
 """
 
 from __future__ import annotations
 
-import asyncio
 import logging
 import time
 from collections.abc import AsyncIterator
-from typing import Any
 
-from app.core.agent_loop.display import render_display_from_map, tool_display_map
-from app.core.observability.workshop import tool_span
+from app.core.agent_loop.events import _prepare_llm_messages
+from app.core.agent_loop.safety import (
+    _check_iteration_safety,
+    _collect_tool_results,
+    _should_stop,
+)
+from app.core.agent_loop.stream_retry import _stream_with_retry, _StreamOutcome
 
 from .types import (
     AgentContext,
@@ -32,169 +48,17 @@ from .types import (
     AgentEvent,
     AgentLoopConfig,
     AgentMessage,
-    AgentSafetyConfig,
     AgentStartEvent,
-    AgentTerminatedEvent,
     AgentTool,
     AssistantMessage,
-    LLMEvent,
     MessageEndEvent,
     MessageStartEvent,
     StreamFn,
-    TextContent,
-    TextDeltaEvent,
-    ThinkingDeltaEvent,
-    ToolCallContent,
-    ToolCallEndEvent,
-    ToolCallStartEvent,
-    ToolResultContent,
-    ToolResultEvent,
-    ToolResultMessage,
     TurnEndEvent,
     TurnStartEvent,
 )
 
 _log = logging.getLogger(__name__)
-
-# Truncation budgets for tool-call trace logs. We log enough to identify a
-# stuck loop ("model keeps calling tool X with args Y") without flooding the
-# log with full tool payloads. Full bodies stay at DEBUG.
-_LOG_ARGS_MAX_CHARS = 500
-_LOG_RESULT_MAX_CHARS = 500
-
-
-def _truncate_for_log(value: Any, max_chars: int) -> str:
-    """Render a tool argument or result for logs, truncating long payloads.
-
-    Args:
-        value: The object to render. Dicts are rendered as ``repr`` so keys
-            stay readable; strings pass through untouched.
-        max_chars: Maximum number of characters in the returned string. The
-            truncation marker counts toward the limit so the returned string
-            is always at most ``max_chars`` characters long.
-
-    Returns:
-        A single-line, length-bounded representation suitable for an INFO
-        log line. Newlines are collapsed so each trace stays on one line.
-    """
-    text = value if isinstance(value, str) else repr(value)
-    text = text.replace("\n", " ").replace("\r", " ")
-    if len(text) <= max_chars:
-        return text
-    return text[: max_chars - 1] + "…"
-
-
-async def _execute_and_log_tool_call(
-    *,
-    tc: Any,
-    tool_map: dict[str, AgentTool],
-    iteration: int,
-    config: AgentLoopConfig,
-) -> tuple[str, bool]:
-    """Execute one tool call, emit ``TOOL_CALL_*`` traces, and return the result.
-
-    Pulled out of the main loop so the per-call observability (args preview,
-    duration, result length) lives in one place and the outer loop stays
-    inside the team's branch/statement budget.
-
-    Args:
-        tc: The ``toolCall`` content block from the assistant message.
-        tool_map: ``name → AgentTool`` lookup built once per loop iteration.
-        iteration: 1-based loop iteration, included in the trace lines so
-            stuck loops are easy to bucket by turn in the log.
-        config: Loop config carrying the optional permission gate; checked
-            before ``tool.execute`` so denied calls become surfaced errors
-            instead of side-effecting executions.
-
-    Returns:
-        A ``(result_text, is_error)`` tuple. ``result_text`` is always
-        non-empty so the caller can hand it straight to ``ToolResultEvent``
-        and ``ToolResultMessage`` without further branching.
-    """
-    name: str = tc["name"]
-    call_id: str = tc["tool_call_id"]
-    arguments: dict[str, Any] = tc["arguments"]
-    args_preview = _truncate_for_log(arguments, _LOG_ARGS_MAX_CHARS)
-    _log.info(
-        "TOOL_CALL_START iteration=%d name=%s tool_call_id=%s args=%s",
-        iteration,
-        name,
-        call_id,
-        args_preview,
-    )
-
-    # Workshop / OTel tool span — gives the live trace viewer a row
-    # per call with the args + result + duration, scoped under the
-    # surrounding ``pawrrtal.turn`` span.  No-op when telemetry off.
-    with tool_span(name=name, tool_call_id=call_id, arguments=arguments) as ts:
-        started_at = time.monotonic()
-        result_text, is_error = await _dispatch_tool_call(
-            tool_map=tool_map,
-            name=name,
-            call_id=call_id,
-            arguments=arguments,
-            config=config,
-        )
-        duration_ms = (time.monotonic() - started_at) * 1000.0
-        ts.record_result(result_text, is_error=is_error)
-
-    _log.info(
-        "TOOL_CALL_RESULT iteration=%d name=%s tool_call_id=%s "
-        "is_error=%s duration_ms=%.1f result_len=%d",
-        iteration,
-        name,
-        call_id,
-        is_error,
-        duration_ms,
-        len(result_text),
-    )
-    _log.debug(
-        "TOOL_CALL_RESULT_BODY tool_call_id=%s body=%s",
-        call_id,
-        _truncate_for_log(result_text, _LOG_RESULT_MAX_CHARS),
-    )
-    return result_text, is_error
-
-
-async def _dispatch_tool_call(
-    *,
-    tool_map: dict[str, AgentTool],
-    name: str,
-    call_id: str,
-    arguments: dict[str, Any],
-    config: AgentLoopConfig,
-) -> tuple[str, bool]:
-    """Resolve permission + execute one tool call, returning ``(result_text, is_error)``.
-
-    Pulled out of :func:`_execute_and_log_tool_call` so the surrounding
-    ``tool_span`` context-manager + log lines keep their own scope and
-    the dispatcher stays inside the team's nesting budget.
-
-    Args:
-        tool_map: ``name → AgentTool`` lookup built once per loop iteration.
-        name: Tool name from the assistant's ``toolCall`` block.
-        call_id: Provider-stable id so a permission gate can echo it.
-        arguments: Kwargs the model produced for the tool.
-        config: Loop config carrying the optional permission gate.
-
-    Returns:
-        ``(result_text, is_error)`` where ``result_text`` is always
-        non-empty.
-    """
-    tool = tool_map.get(name)
-    if tool is None:
-        return f"Tool '{name}' not found.", True
-    permission_denial = await _check_permission_or_none(
-        config=config,
-        tool_name=name,
-        arguments=arguments,
-    )
-    if permission_denial is not None:
-        return permission_denial, True
-    try:
-        return await tool.execute(call_id, **arguments), False
-    except Exception as exc:
-        return f"Tool error: {exc}", True
 
 
 async def agent_loop(
@@ -327,501 +191,3 @@ async def _run_loop(
             break
 
     yield AgentEndEvent(type="agent_end", messages=new_messages)
-
-
-def _check_iteration_safety(
-    safety: AgentSafetyConfig,
-    iteration: int,
-    started_at: float,
-) -> AgentTerminatedEvent | None:
-    """Return a termination event if any pre-turn safety limit is hit, else None."""
-    if safety.max_iterations is not None and iteration >= safety.max_iterations:
-        return _terminated(
-            reason="max_iterations",
-            message=(
-                f"Agent stopped: hit max_iterations cap of "
-                f"{safety.max_iterations}.  This usually means the "
-                "model got stuck in a tool-call loop.  Reply with "
-                "new context or raise the cap if the work needs "
-                "more steps."
-            ),
-            limit=safety.max_iterations,
-            observed=iteration,
-        )
-
-    elapsed = time.monotonic() - started_at
-    if safety.max_wall_clock_seconds is not None and elapsed >= safety.max_wall_clock_seconds:
-        return _terminated(
-            reason="max_wall_clock",
-            message=(
-                f"Agent stopped: hit wall-clock budget of "
-                f"{safety.max_wall_clock_seconds:.0f}s.  Raise "
-                "`max_wall_clock_seconds` for legitimately long "
-                "turns."
-            ),
-            limit_seconds=safety.max_wall_clock_seconds,
-            observed_seconds=round(elapsed, 2),
-            iterations=iteration,
-        )
-
-    return None
-
-
-async def _prepare_llm_messages(
-    messages: list[AgentMessage],
-    config: AgentLoopConfig,
-) -> list[AgentMessage]:
-    """Apply context transform (if any) and convert to LLM format."""
-    transformed = messages
-    if config.transform_context is not None:
-        transformed = await config.transform_context(list(messages))
-    return config.convert_to_llm(transformed)
-
-
-async def _collect_tool_results(
-    tool_calls: list[Any],
-    tool_map: dict[str, AgentTool],
-    messages: list[AgentMessage],
-    new_messages: list[AgentMessage],
-    safety: AgentSafetyConfig,
-    consecutive_tool_errors: int,
-    iteration: int,
-    config: AgentLoopConfig,
-) -> tuple[list[ToolResultEvent], list[ToolResultMessage], int, AgentTerminatedEvent | None]:
-    """Execute all tool calls for one turn and collect result events.
-
-    Returns ``(result_events, tool_results, consecutive_tool_errors, terminated_event)``.
-    ``terminated_event`` is non-None only when the consecutive-error budget is hit.
-    Messages are appended to both ``messages`` and ``new_messages`` in-place.
-    """
-    result_events: list[ToolResultEvent] = []
-    tool_results: list[ToolResultMessage] = []
-    terminated: AgentTerminatedEvent | None = None
-
-    for tc in tool_calls:
-        if tc["type"] != "toolCall":
-            continue
-        result_text, is_error = await _execute_and_log_tool_call(
-            tc=tc,
-            tool_map=tool_map,
-            iteration=iteration,
-            config=config,
-        )
-        result_events.append(
-            ToolResultEvent(
-                type="tool_result",
-                tool_call_id=tc["tool_call_id"],
-                content=result_text,
-                is_error=is_error,
-            )
-        )
-        tool_result_msg = ToolResultMessage(
-            role="toolResult",
-            tool_call_id=tc["tool_call_id"],
-            name=tc["name"],
-            content=[ToolResultContent(type="text", text=result_text)],
-            is_error=is_error,
-        )
-        tool_results.append(tool_result_msg)
-        messages.append(tool_result_msg)
-        new_messages.append(tool_result_msg)
-
-        if is_error:
-            consecutive_tool_errors += 1
-            if (
-                safety.max_consecutive_tool_errors is not None
-                and consecutive_tool_errors >= safety.max_consecutive_tool_errors
-            ):
-                terminated = _terminated(
-                    reason="consecutive_tool_errors",
-                    message=(
-                        "Agent stopped: "
-                        f"{consecutive_tool_errors} tool calls "
-                        "failed back-to-back.  The model is "
-                        "likely retrying a broken tool with the "
-                        "same arguments.  Inspect the last tool "
-                        "errors and either fix the inputs or "
-                        "raise `max_consecutive_tool_errors`."
-                    ),
-                    limit=safety.max_consecutive_tool_errors,
-                    observed=consecutive_tool_errors,
-                    iterations=iteration,
-                )
-                break
-        else:
-            consecutive_tool_errors = 0
-
-    return result_events, tool_results, consecutive_tool_errors, terminated
-
-
-def _should_stop(
-    stop_reason: str,
-    tool_calls: list[Any],
-    config: AgentLoopConfig,
-    messages: list[AgentMessage],
-    tools: list[AgentTool],
-) -> bool:
-    """Return True if the loop should halt after this turn."""
-    if stop_reason in {"error", "aborted"}:
-        return True
-    if config.should_stop_after_turn is not None:
-        fake_ctx = _make_context_snapshot(messages, tools)
-        if config.should_stop_after_turn(fake_ctx):
-            return True
-    return not tool_calls
-
-
-# ---------------------------------------------------------------------------
-# Stream-with-retry helper
-# ---------------------------------------------------------------------------
-
-
-class _StreamOutcome:
-    """Result of one (possibly retried) stream attempt.
-
-    Either the stream succeeded — in which case ``events`` holds the
-    AgentEvents we want the caller to yield in order, ``assistant_content``
-    + ``stop_reason`` carry the final assistant message, and
-    ``terminated_event`` is ``None`` — or every retry was exhausted, in
-    which case ``terminated_event`` carries the safety termination notice
-    and the other fields are empty defaults.
-    """
-
-    __slots__ = (
-        "assistant_content",
-        "consecutive_llm_errors_after",
-        "events",
-        "provider_state",
-        "stop_reason",
-        "terminated_event",
-    )
-
-    def __init__(
-        self,
-        events: list[AgentEvent],
-        assistant_content: list[TextContent | ToolCallContent],
-        stop_reason: str,
-        consecutive_llm_errors_after: int,
-        terminated_event: AgentTerminatedEvent | None,
-        provider_state: dict[str, Any] | None = None,
-    ) -> None:
-        self.events = events
-        self.assistant_content = assistant_content
-        self.stop_reason = stop_reason
-        self.consecutive_llm_errors_after = consecutive_llm_errors_after
-        self.terminated_event = terminated_event
-        self.provider_state = provider_state
-
-
-class _StreamAttemptState:
-    """Mutable state for one provider stream attempt."""
-
-    __slots__ = ("emitted_event", "provider_state")
-
-    def __init__(self) -> None:
-        self.emitted_event = False
-        self.provider_state: dict[str, Any] | None = None
-
-
-async def _stream_with_retry(
-    stream_fn: StreamFn,
-    llm_messages: list[AgentMessage],
-    tools: list[AgentTool],
-    safety: AgentSafetyConfig,
-    consecutive_llm_errors: int,
-    outcome: _StreamOutcome,
-) -> AsyncIterator[AgentEvent]:
-    """Stream one assistant turn, updating ``outcome`` with final state."""
-    backoff = max(safety.llm_retry_backoff_seconds, 0.0)
-    max_errors = safety.max_consecutive_llm_errors
-    attempts = 0
-    display_by_name = tool_display_map(tools)
-
-    while True:
-        attempts += 1
-        attempt_state = _StreamAttemptState()
-
-        try:
-            async for event in _stream_attempt_events(
-                stream_fn(llm_messages, tools),
-                outcome,
-                attempt_state,
-                display_by_name,
-            ):
-                yield event
-        except Exception as exc:
-            if attempt_state.emitted_event:
-                outcome.terminated_event = _stream_interrupted_after_events(exc)
-                yield outcome.terminated_event
-                return
-
-            consecutive_llm_errors += 1
-            _log.warning(
-                "agent_loop: provider stream failed (attempt %d, consecutive=%d/%s): %s",
-                attempts,
-                consecutive_llm_errors,
-                max_errors if max_errors is not None else "∞",
-                exc,
-            )
-            exhausted = _retry_budget_exhausted(
-                max_errors=max_errors,
-                consecutive_llm_errors=consecutive_llm_errors,
-                exc=exc,
-            )
-            if exhausted is not None:
-                _copy_stream_outcome(exhausted, outcome)
-                assert outcome.terminated_event is not None
-                yield outcome.terminated_event
-                return
-
-            await _sleep_before_retry(backoff, attempts)
-            continue
-
-        outcome.consecutive_llm_errors_after = 0
-        outcome.provider_state = attempt_state.provider_state
-        return
-
-
-async def _sleep_before_retry(backoff: float, attempts: int) -> None:
-    """Sleep for exponential retry backoff when configured."""
-    if backoff <= 0:
-        return
-    wait = min(backoff * (2 ** (attempts - 1)), 30.0)
-    await asyncio.sleep(wait)
-
-
-async def _stream_attempt_events(
-    stream: AsyncIterator[LLMEvent],
-    outcome: _StreamOutcome,
-    attempt_state: _StreamAttemptState,
-    display_by_name: dict[str, Any],
-) -> AsyncIterator[AgentEvent]:
-    """Yield translated events for one provider stream attempt."""
-    async for llm_event in stream:
-        events: list[AgentEvent] = []
-        done = _consume_llm_event(llm_event, events, display_by_name)
-        if events:
-            attempt_state.emitted_event = True
-            outcome.events.extend(events)
-        _apply_done_event(done, outcome, attempt_state)
-        for event in events:
-            yield event
-
-
-def _apply_done_event(
-    done: dict[str, Any] | None,
-    outcome: _StreamOutcome,
-    attempt_state: _StreamAttemptState,
-) -> None:
-    """Capture the final provider state from a terminal ``done`` event."""
-    if done is None:
-        return
-    outcome.assistant_content = done["content"]
-    outcome.stop_reason = done["stop_reason"]
-    attempt_state.provider_state = _carry_provider_state(done, attempt_state.provider_state)
-
-
-def _retry_budget_exhausted(
-    *,
-    max_errors: int | None,
-    consecutive_llm_errors: int,
-    exc: Exception,
-) -> _StreamOutcome | None:
-    """Return a terminating :class:`_StreamOutcome` if retry budget exhausted.
-
-    Pulled out of :func:`_stream_with_retry` so the inner loop stays
-    within the project's nesting-depth budget (depth 3) — enforced by
-    ``scripts/check-nesting.py``.  Returns ``None`` when the budget
-    still has room and the caller should retry.
-    """
-    if max_errors is None or consecutive_llm_errors < max_errors:
-        return None
-    terminated = _terminated(
-        reason="consecutive_llm_errors",
-        message=(
-            "Agent stopped: "
-            f"{consecutive_llm_errors} provider errors in a "
-            "row.  The upstream model is likely down or "
-            "rate-limiting.  Try again, switch model, or "
-            "raise `max_consecutive_llm_errors`.  Last "
-            f"error: {exc}"
-        ),
-        limit=max_errors,
-        observed=consecutive_llm_errors,
-        last_error=str(exc),
-    )
-    return _StreamOutcome(
-        events=[],
-        assistant_content=[],
-        stop_reason="error",
-        consecutive_llm_errors_after=consecutive_llm_errors,
-        terminated_event=terminated,
-    )
-
-
-def _copy_stream_outcome(source: _StreamOutcome, target: _StreamOutcome) -> None:
-    """Copy final stream state into the caller-owned outcome object."""
-    target.events = source.events
-    target.assistant_content = source.assistant_content
-    target.stop_reason = source.stop_reason
-    target.consecutive_llm_errors_after = source.consecutive_llm_errors_after
-    target.terminated_event = source.terminated_event
-    target.provider_state = source.provider_state
-
-
-def _stream_interrupted_after_events(exc: Exception) -> AgentTerminatedEvent:
-    """Build the no-retry termination used after partial output was emitted."""
-    return _terminated(
-        reason="stream_interrupted_after_events",
-        message=(
-            "Agent stopped: provider stream failed after partial output was "
-            f"already emitted. Last error: {exc}"
-        ),
-        last_error=str(exc),
-    )
-
-
-def _consume_llm_event(
-    llm_event: LLMEvent,
-    events: list[AgentEvent],
-    display_by_name: dict[str, Any] | None = None,
-) -> dict[str, Any] | None:
-    """Translate one LLMEvent into AgentEvents and append to ``events``.
-
-    Returns the payload for the terminal ``done`` event
-    (``{content, stop_reason}``) when the consumed event was that
-    terminator; otherwise ``None``.  Returning instead of using a
-    mutable out-param keeps the caller's loop body flat enough to fit
-    inside the project nesting-depth budget.
-    """
-    if llm_event["type"] == "text_delta":
-        events.append(TextDeltaEvent(type="text_delta", text=llm_event["text"]))
-        return None
-    if llm_event["type"] == "thinking_delta":
-        events.append(ThinkingDeltaEvent(type="thinking_delta", text=llm_event["text"]))
-        return None
-    if llm_event["type"] == "tool_call":
-        display = render_display_from_map(
-            display_by_name or {},
-            llm_event["name"],
-            llm_event["arguments"],
-        )
-        events.append(
-            ToolCallStartEvent(
-                type="tool_call_start",
-                tool_call_id=llm_event["tool_call_id"],
-                name=llm_event["name"],
-            )
-        )
-        events.append(
-            ToolCallEndEvent(
-                type="tool_call_end",
-                tool_call_id=llm_event["tool_call_id"],
-                name=llm_event["name"],
-                arguments=llm_event["arguments"],
-                display=display,
-            )
-        )
-        return None
-    if llm_event["type"] == "done":
-        result: dict[str, Any] = {
-            "content": llm_event["content"],
-            "stop_reason": llm_event["stop_reason"],
-        }
-        # Forward opaque provider replay state if the StreamFn populated
-        # it.  ``provider_state`` is a ``total=False`` field on
-        # ``LLMDoneEvent`` so we read it with ``.get`` and only copy it
-        # forward when the provider actually returned one.  See
-        # ``_stream_with_retry`` for the per-turn capture.
-        provider_state = llm_event.get("provider_state")
-        if provider_state is not None:
-            result["provider_state"] = provider_state
-        return result
-    return None
-
-
-def _carry_provider_state(
-    done: dict[str, Any] | None,
-    current: dict[str, Any] | None,
-) -> dict[str, Any] | None:
-    """Return the provider_state from ``done`` if present, else keep ``current``.
-
-    Extracted out of :func:`_stream_with_retry` so the streaming loop's
-    body stays under the project nesting-depth budget.  The loop never
-    inspects the contents — providers own the keyspace; we just forward
-    the slot.
-    """
-    if done is not None and "provider_state" in done:
-        state = done["provider_state"]
-        # ``done`` is a TypedDict-shaped dict — narrow the dynamic value
-        # back to the declared return type so mypy doesn't infer ``Any``.
-        return state if isinstance(state, dict) else None
-    return current
-
-
-def _terminated(
-    *,
-    reason: str,
-    message: str,
-    **details: Any,
-) -> AgentTerminatedEvent:
-    """Build an :class:`AgentTerminatedEvent` with structured details.
-
-    Centralised so every termination path uses the same dict shape.
-    """
-    return AgentTerminatedEvent(
-        type="agent_terminated",
-        reason=reason,  # type: ignore[typeddict-item]
-        details=details,
-        message=message,
-    )
-
-
-def _make_context_snapshot(
-    messages: list[AgentMessage],
-    tools: list[AgentTool],
-) -> AgentContext:
-    """Build an AgentContext snapshot for the should_stop_after_turn predicate."""
-    return AgentContext(system_prompt="", messages=messages, tools=tools)
-
-
-async def _check_permission_or_none(
-    *,
-    config: AgentLoopConfig,
-    tool_name: str,
-    arguments: dict[str, Any],
-) -> str | None:
-    """Run the configured permission gate, returning a denial message.
-
-    Returns ``None`` when there's no gate or when the gate allows the
-    call. Otherwise returns the denial reason as a string so the
-    caller can surface it as the tool result. Also fires the optional
-    ``permission_audit_sink`` (errors swallowed — audit failures must
-    never break a turn).
-
-    Kept as a module-level helper so the inner loop body in
-    :func:`_run_loop` stays under the project's nesting budget.
-    """
-    if config.permission_check is None:
-        return None
-    try:
-        decision = await config.permission_check(tool_name, arguments)
-    except Exception as exc:
-        # A crashed permission check is a configuration bug, not a
-        # security signal — fail closed (deny) so a broken policy
-        # doesn't silently allow tool use, but include the error so
-        # the operator notices in logs.
-        _log.exception("agent_loop: permission_check crashed; failing closed for %s", tool_name)
-        return f"Tool '{tool_name}' denied: permission check error ({exc})."
-
-    if decision.get("allow", False):
-        return None
-
-    reason = decision.get("reason") or "Tool call denied by permission policy."
-    if config.permission_audit_sink is not None:
-        try:
-            await config.permission_audit_sink(tool_name, arguments, decision)
-        except Exception:
-            # Swallow audit failures — never break a turn over them.
-            _log.exception("agent_loop: permission_audit_sink raised; ignoring")
-    return reason
